app/requirements.txt
Pillow
pytesseract
uvicorn
gunicorn
fastapi[all]
wheel
python-multipart
jinja2
numpy
opencv-python
python-dotenv
numpy
requests
ocrmypdf
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pypdf
minio
boto3
python-decouple